:sparkles: feat: Real time chat
There are more changes than just the integration of socket io, such as
a really important change on the model regarding the structure of chat
that rippled through other parts of the model.
This happened because the initial integration with socket was already
done prior to my realization that something need to be done about the
whole model.